add auth service stub
